新增流水表和接口

+++ lx-admin/src/main/java/io/lx/modules/wxapp/entity/TbOrdersEntity.java
@@ -35,7 +35,7 @@ public class TbOrdersEntity {
     /**
      * 
      */
-	private Long productId;
+	private Integer productId;
     /**
      * 
      */
